support ipex2.1
