+++ neural_compressor/adaptor/torch_utils/util.py
@@ -17,6 +17,7 @@
 """Util Class and Functions."""
 import copy
 import re
+import json
 import numpy as np
 from collections import UserDict
 from packaging.version import Version
@@ -493,6 +494,40 @@ def get_quantizable_ops_from_cfgs(ops_name, op_infos_from_cfgs, input_tensor_ids
             for q_op in q_ops:
                 quantizable_ops.append(q_op)
     return quantizable_ops
+
+def update_sq_scale(ipex_config_path, smoothquant_scale_info):
+    """update ipex_config.json with smoothquant scale info generated by our algorithm.
+
+    Args:
+        ipex_config_path (str): a path to temporary ipex_config.json file.
+        smoothquant_scale_info (dict): a dict contains smoothquant scale info.
+    """
+    with open(ipex_config_path, 'r') as f:
+        ipex_config = json.load(f)
+        for module_name, v in ipex_config.items():
+            if 'q_op_infos' in v and v['q_op_infos']:
+                for op_num, v1 in v['q_op_infos'].items():
+                    # update alpha data instead of updating weight scale
+                    op_name = v1['fqn'] # fqn always exists even it's empty.
+                    if op_name in smoothquant_scale_info:
+                        # observers were overridden by the fallback step, setting it back.
+                        v1['activation_observer'] = {'name': 'SmoothQuantActivationObserver', 
+                                        'smooth_quant_enabled': False, 'dtype': 'torch.quint8', 
+                                        'qscheme': 'torch.per_tensor_affine', 'reduce_range': False,
+                                        'quant_min': 0, 'quant_max': 255, 
+                                        'alpha': smoothquant_scale_info[op_name]['alpha']
+                                        }
+                        v1['weight_observer'] = {'name': 'SmoothQuantWeightObserver', 
+                                        'smooth_quant_enabled': False, 'dtype': 'torch.qint8', 
+                                        'qscheme': 'torch.per_channel_symmetric', 'reduce_range': False, 
+                                        'quant_min': -128, 'quant_max': 127, 
+                                        'alpha': smoothquant_scale_info[op_name]['alpha'] #only update alpha
+                                        }
+        f.close()
+    # overwrite ipex_config_path
+    with open(ipex_config_path, 'w') as f1:
+        json.dump(ipex_config, f1, indent = 4)
+        f1.close()
 
 def auto_copy(module):  # pragma: no cover
     """Get an IPEX prepared model and return a fp32 model.
